requirements.txt
appdirs==1.4.0
attrs==16.3.0
beautifulsoup4==4.5.3
bs4==0.0.1
cffi==1.9.1
chardet==2.3.0
click==6.7
constantly==15.1.0
cryptography==1.7.2
cssselect==1.0.1
dominate==2.3.1
EasyProcess==0.2.3
Flask==0.12
Flask-Bootstrap==3.3.7.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
get==0.0.20
idna==2.2
incremental==16.10.1
itsdangerous==0.24
Jinja2==2.9.5
lxml==3.7.2
MarkupSafe==0.23
olefile==0.44
packaging==16.8
parsel==1.1.0
pefile==2016.3.28
Pillow==4.0.0
post==0.0.12
public==0.0.37
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
PyExecJS==1.4.0
Pygments==2.2.0
pyinstaller==3.3.dev0+483c819
pyOpenSSL==16.2.0
pyparsing==2.1.10
PyQt5==5.7.1
pyqt5-tools==5.7.dev9
PySocks==1.6.6
pytesseract==0.1.6
python-nmap==0.6.1
pywin32==220
query-string==0.0.11
queuelib==1.4.2
request==0.0.12
requests==2.13.0
rsa==3.4.2
scapy-python3==0.20
Scrapy==1.3.0
selenium==3.0.2
service-identity==16.0.0
setupfiles==0.0.0
simplejson==3.10.0
sip==4.19
six==1.10.0
SQLAlchemy==1.2.4
tea==0.0.5
Twisted==16.6.0
virtualenv==15.1.0
visitor==0.1.3
w3lib==1.16.0
Werkzeug==0.11.15
WTForms==2.1
xlrd==1.0.0
xlutils==2.0.0
xlwt==1.2.0
zope.interface==4.3.3
pymysql==0.8.0